Refactors helpers to clientHelpers and moves uiHelpers to clientHelpers
