+++ components/forms/Form/Form.test.js
@@ -2,11 +2,11 @@ import React from "react";
 import { cleanup, render, screen, waitFor, fireEvent } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { Form } from "@components/forms";
-import { submitToAPI } from "@lib/helpers";
+import { submitToAPI } from "@lib/clientHelpers";
 import { useFlag } from "@lib/hooks/useFlag";
 
-jest.mock("@lib/helpers", () => {
-  const originalModule = jest.requireActual("@lib/helpers");
+jest.mock("@lib/clientHelpers", () => {
+  const originalModule = jest.requireActual("@lib/clientHelpers");
   return {
     __esModule: true,
     ...originalModule,
