+++ lib/tests/clientHelpers.test.js
@@ -1,5 +1,5 @@
 import mockedAxios from "axios";
-import { buildFormDataObject, rehydrateFormResponses, submitToAPI } from "../helpers";
+import { buildFormDataObject, rehydrateFormResponses, submitToAPI } from "../clientHelpers";
 import { logMessage } from "@lib/logger";
 import TestForm from "../../__fixtures__/testData.json";
 import { getCsrfToken } from "next-auth/react";
