+++ lib/tests/csrfProtected.test.js
@@ -1,6 +1,6 @@
 import mockedAxios from "axios";
 import { createMocks } from "node-mocks-http";
-import { submitToAPI } from "../helpers";
+import { submitToAPI } from "../clientHelpers";
 import { getCsrfToken } from "next-auth/react";
 import { csrfProtected } from "@lib/middleware/csrfProtected";
 import TestForm from "../../__fixtures__/testData.json";
